Finished Rest Api for products. Finished Tests

+++ Complevo.Assesment.Data/Migrations/ProductContextModelSnapshot.cs
@@ -23,11 +23,11 @@ namespace Complevo.Assesment.Data.Migrations
 
             modelBuilder.Entity("Complevo.Assesment.Data.Entities.Product", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("int");
+                        .HasColumnType("bigint");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"), 1L, 1);
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"), 1L, 1);
 
                     b.Property<string>("Description")
                         .IsRequired()
